net: openthread: rpc: restructure server code
Split OpenThread RPC server functions into source files
that correspond with the OpenThread code structure.
