+++ subsys/net/openthread/rpc/client/ot_rpc_thread.c
