+++ subsys/net/openthread/rpc/server/ot_rpc_cli.c
