+++ subsys/net/openthread/rpc/server/ot_rpc_diag.c
