+++ subsys/net/openthread/rpc/server/ot_rpc_message.h
