+++ subsys/net/openthread/rpc/server/ot_rpc_udp.c
@@ -4,6 +4,8 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
  */
 
+#include "ot_rpc_message.h"
+
 #include <nrf_rpc/nrf_rpc_serialize.h>
 #include <ot_rpc_ids.h>
 #include <ot_rpc_types.h>
@@ -11,16 +13,9 @@
 
 #include <nrf_rpc_cbor.h>
 
-#include <openthread/error.h>
-#include <openthread/instance.h>
-
 #include <openthread/udp.h>
 
 #include <zephyr/net/openthread.h>
-#include <zcbor_decode.h>
-#include <string.h>
-
-#include "ot_rpc_message_server.h"
 
 #define OT_UDP_SOCKETS 8
